continue working on setup script

+++ app/utils.py
@@ -15,20 +15,21 @@ HOOKS_DIR = DATA_DIR / 'hooks'
 HOOKS_DIR.mkdir(exist_ok=True)
 WORKDIR = DATA_DIR / 'workdir'
 WORKDIR.mkdir(exist_ok=True)
-CONFIG_FILENAME = BASE_DIR / 'config.yaml'
+CONFIG_FILENAME = DATA_DIR / 'config.yml'
+SESSION_FILENAME = DATA_DIR / 'TGPy.session'
 
 
 def migrate_from_old_versions():
     from app.app_config import Config
 
     old_session_file = BASE_DIR / 'TGPy.session'
-    if old_session_file.exists():
-        old_session_file.rename(DATA_DIR / 'TGPy.session')
+    if old_session_file.exists() and not SESSION_FILENAME.exists():
+        old_session_file.rename(SESSION_FILENAME)
     old_config_file = BASE_DIR / 'config.py'
-    if old_config_file.exists():
+    if old_config_file.exists() and not CONFIG_FILENAME.exists():
         try:
             config_mod = importlib.import_module('config')
-            config = Config(api_id=config_mod.api_id, api_hash=config_mod.api_hash, phone=config_mod.phone)
+            config = Config(api_id=config_mod.api_id, api_hash=config_mod.api_hash)
             with open(CONFIG_FILENAME, 'w') as file:
                 yaml.safe_dump(config.dict(), file)
         except (ValidationError, AttributeError, ImportError):
@@ -87,5 +88,5 @@ def get_version():
     return 'unknown'
 
 
-__all__ = ['BASE_DIR', 'DATA_DIR', 'HOOKS_DIR', 'WORKDIR', 'CONFIG_FILENAME', 'yaml_multiline_str', 'run_cmd',
-           'get_version', 'migrate_from_old_versions']
+__all__ = ['BASE_DIR', 'DATA_DIR', 'HOOKS_DIR', 'WORKDIR', 'CONFIG_FILENAME', 'SESSION_FILENAME',
+           'yaml_multiline_str', 'run_cmd', 'get_version', 'migrate_from_old_versions']
